+++ frontend/src/routes/channels/id/FetchVideosAction.tsx
@@ -1,6 +1,6 @@
 import {fetchedVideos} from "../../../store/slices/fetches";
 import {useAppDispatch} from "../../../store/hooks";
-import {useFetchVideosByChannelIdMutation} from "../../../store/slices/api";
+import {apiUtils, useFetchVideosByChannelIdMutation, videosAdapter} from "../../../store/slices/api";
 import {NotificationsSource} from "../../../components/NotificationsSource";
 import {EventSourceProvider} from "react-sse-hooks";
 import {FetchVideosEvent} from "../../../model/events/FetchEvent";
@@ -21,6 +21,11 @@ export const FetchVideosAction = ({channelId, mode}: FetchVideosWidgetProps) => 
 
     const handleEvent = (event: FetchVideosEvent) => {
         dispatch(fetchedVideos(event));
+        if (event.status !== 'FAILED') {
+            dispatch(apiUtils.updateQueryData('listVideosByChannelId', channelId, (draft) => {
+                videosAdapter.setMany(draft, event.items)
+            }))
+        }
     }
 
     return (