+++ frontend/src/routes/channels/id/FetchVideosAlert.tsx
@@ -70,8 +70,8 @@ const FetchAvailableBody = ({channel}: FetchVideosAlertProps) => {
 
 const FetchRequestedBody = ({channel, fetchState}: AlertBodyProps) => {
 
-    const videoCountLabel = fetchState.data.length === 0
-        ? "latest videos" : pluralize(fetchState.data.length, "video");
+    const videoCountLabel = (fetchState.data?.length ?? 0) === 0
+        ? "latest videos" : pluralize(fetchState.data?.length ?? 0, "video");
 
     return (
         <Alert className={"d-flex align-items-center alert-secondary"}>
@@ -91,7 +91,7 @@ const FetchRequestedBody = ({channel, fetchState}: AlertBodyProps) => {
 const FetchCompletedBody = ({fetchState}: AlertBodyProps) => (
     <Alert className={"d-flex align-items-center alert-success"}>
         <Col>
-            Fetched <strong>{pluralize(fetchState.data.length, "video")}</strong> from YouTube.
+            Fetched <strong>{pluralize(fetchState.data?.length ?? 0, "video")}</strong> from YouTube.
         </Col>
         <Col xs={"auto"}>
             <Button className={"d-flex align-items-center btn-outline-success"} disabled={true}>