+++ frontend/src/routes/videos/id/FetchCommentsAction.tsx
@@ -1,4 +1,4 @@
-import {useFetchCommentsByVideoIdMutation} from "../../../store/slices/api";
+import {apiUtils, commentsAdapter, useFetchCommentsByVideoIdMutation} from "../../../store/slices/api";
 import {useAppDispatch} from "../../../store/hooks";
 import {EventSourceProvider} from "react-sse-hooks";
 import {NotificationsSource} from "../../../components/NotificationsSource";
@@ -20,6 +20,11 @@ export const FetchCommentsAction = ({videoId}: FetchVideosActionProps) => {
 
     const handleEvent = (event: FetchCommentsEvent) => {
         dispatch(fetchedComments(event));
+        if (event.status !== 'FAILED') {
+            dispatch(apiUtils.updateQueryData("listCommentsByVideoId", videoId, draft => {
+                commentsAdapter.addMany(draft, event.items);
+            }))
+        }
     }
 
     return (